+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* Add a 5 second timeout to public API calls via new `Config.timeoutMS` param
 * Implement a reload based on time of most recent toot in timeline (10 minutes is the trigger)
 * Get rid of unnecessary `Toot.resolveAttempted` flag